--- .github/ISSUE_TEMPLATE/blank-issue.md
@@ -1,13 +1,0 @@
----
-
-name: 'Blank Issue'
-about: 'This template is used to create a blank issue without a template. Please use our templates where applicable.'
-title: 'The title of your issue.'
-ref: 'main'
-labels:
-  - 'State: Conversation'
-
----
-
-We are assuming that you are an expert user. Please follow our templates unless
-you know what you are doing.
