--- .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,31 +1,0 @@
----
-
-name: 'Bug Report'
-about: 'This template is used to report unintended behavior of our projects.'
-title: '[BUG] A brief overview of your concern.'
-ref: 'main'
-labels:
-  - 'Type: Bug'
-  - 'State: Conversation'
-
----
-
-**Describe the Bug**: A clear and concise description of what is problematic.
-
-**Information / Steps to Reproduce**:
-
-1. Provide detailed steps on how to reproduce the issue.
-2. Key aspects of how the environment was set up.
-3. Code or repository that can be used to consistently reproduce the behavior.
-4. Please include any and all errors that are generated.
-
-**Expected Behavior**: Please explain the behavior you expected the application
-to implement.
-
-**Current Environment**: Please detail the specs and operating system of the
-machine you are currently using. If possible, please replicate this issue across
-multiple different operating systems. Please include the version of our project
-that you are using.
-
-**Additional Context**: Please provide any additional information you feel
-should be shared, if applicable.
