--- .github/ISSUE_TEMPLATE/enhancement-request.md
@@ -1,28 +1,0 @@
----
-
-name: 'Enhancement Request'
-about: 'This template is used to request enhancements to be added in our projects.'
-title: '[FEAT] A brief overview of your request.'
-ref: 'main'
-labels:
-  - 'Type: Enhancement'
-  - 'State: Conversation'
-
----
-
-**Describe the Request**: A clear and concise description of what you are
-requesting.
-
-**Information**:
-
-1. How would you like to see this implemented?
-2. How do you feel this will benefit the project?
-3. Is there any existing functionality that may need changed to support this?
-4. Will this functionality replace or otherwise deprecate existing
-   functionality?
-
-**Expected Behavior**: Please explain the behavior you are expecting the
-application to implement.
-
-**Additional Context**: Please provide any additional information you feel
-should be shared, if applicable.
