--- .github/ISSUE_TEMPLATE/security-concern.md
@@ -1,31 +1,0 @@
----
-
-name: 'Security Concern'
-about: 'This template is used to report concerns over the security implementations in our projects.'
-title: '[SEC] A brief overview of your concern.'
-ref: 'main'
-labels:
-  - 'Type: Infrastructure'
-  - 'Priority: Critical'
-  - 'State: Conversation'
-
----
-
-**Describe the Concern**: A clear and concise description of what is
-problematic.
-
-**Information / Steps to Reproduce**:
-
-1. Provide detailed steps on how to reproduce the issue.
-2. Key aspects of how the environment was set up.
-3. Code or repository that can be used to consistently reproduce the behavior.
-4. Please include any and all errors that are generated.
-
-**Expected Behavior**: Please explain the behavior you expected the application
-to implement.
-
-**Mitigation**: Please explain how you believe that we can mitigate this concern
-to prevent exploitation.
-
-**Additional Context**: Please provide any additional information you feel
-should be shared, if applicable.
