--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,49 +1,0 @@
-# Pull Request
-
-## Summary and Information
-
-<!--
-Please include a summary of the changes and which issue is related. Reference
-your motivation and context to the issue. Please include additions or changes to
-dependencies for security analysis by other contributors.
-
-Fill in the issues related to this issue in their respective category.
--->
-
-...
-
-Referencing Issue(s): #
-
-Fixing Issue(s): #
-
-## Summary of Testing
-
-To what extent did you test this pull request? Did any tests require
-modification to result in a passing state (breaking changes)?
-
-I tested this PR by ... (description of testing methods)
-
-- [ ] Tests passed with no modification.
-- [ ] Tests passed with modification.
-- [ ] Tests are not currently passing.
-- [ ] Tests need to be added to exist for the case covered in this issue.
-
-## Checklist for Integrity
-
-- [ ] The changes follow the style guidelines defined in the project.
-- [ ] I have performed a self-review of my code, including testing intentionally
-  incorrect information.
-- [ ] I have commented my code, particularly in hard to understand or complex
-  areas.
-- [ ] The changes have not generated any new errors or warning by
-  static-analysis tooling.
-- [ ] All dependant changes have been merged and published in downstream
-  modules.
-- [ ] To the best of my knowledge, all information in this pull request is
-  vulnerability free and well researched.
-    - [ ] This pull request may require more research or analysis of complex logic
-      or technology.
-    - [ ] This pull request contains regular expression addition or modifications
-      which may need verification.
-    - [ ] This pull request contains database logic which may need verification of
-      proper user-input sterilization.
