--- .github/settings.yml
@@ -1,85 +1,0 @@
-repository:
-    security_and_analysis:
-        advanced_security:
-            status: enabled
-        secret_scanning:
-            status: enabled
-
-    has_issues: true
-    has_projects: true
-    has_wiki: true
-    default_branch: main
-
-    allow_squash_merge: true
-    allow_merge_commit: true
-    allow_rebase_merge: true
-    allow_auto_merge: false
-    delete_branch_on_merge: true
-
-    enable_automated_security_fixes: true
-    enable_vulnerability_alerts: true
-
-labels:
-    - name: "Complexity: High"
-      color: "F9D0C4"
-    - name: "Complexity: Medium"
-      color: "FEF2C0"
-    - name: "Complexity: Low"
-      color: "D4C5F9"
-    - name: "Priority: Critical"
-      color: "000000"
-    - name: "Priority: High"
-      color: "F9D0C4"
-    - name: "Priority: Medium"
-      color: "FEF2C0"
-    - name: "Priority: Low"
-      color: "D4C5F9"
-    - name: "State: Help Wanted"
-      color: "0E8A16"
-    - name: "State: Work in Progress"
-      color: "5319E7"
-    - name: "State: Conversation"
-      color: "27828C"
-    - name: "State: Blocked"
-      color: "000000"
-    - name: "State: Postponed"
-      color: "C5DEF5"
-    - name: "State: Unassigned"
-      color: "BFDADC"
-    - name: "Type: Enhancement"
-      color: "1D76DB"
-    - name: "Type: Bug"
-      color: "B60205"
-    - name: "Type: Infrastructure"
-      color: "FBCA04"
-    - name: "Type: Dependencies"
-      color: "D4C5F9"
-    - name: "Meta: Question"
-      color: "BFDADC"
-    - name: "Meta: Duplicate"
-      color: "BFDADC"
-    - name: "Meta: Abandoned"
-      color: "808080"
-
-branches:
-    - name: main
-      protection:
-          required_pull_request_reviews:
-              dismiss_stale_reviews: true
-              require_code_owner_reviews: true
-              required_approving_review_count: 1
-              dismissal_restrictions:
-                  users: []
-                  teams: []
-          required_status_checks:
-              strict: true
-              contexts: []
-          enforce_admins: false
-          required_linear_history: true
-          allow_force_pushes: false
-          allow_deletions: false
-          required_conversation_resolution: true
-          restrictions:
-              apps: []
-              users: []
-              teams: []
