--- .github/stale.yml
@@ -1,23 +1,0 @@
-daysUntilStale: 14
-daysUntilClose: 3
-exemptLabels:
-    - "Priority: Critical"
-    - "Priority: High"
-    - "Priority: Medium"
-    - "Priority: Low"
-    - "State: Help Wanted"
-    - "State: Postponed"
-    - "State: Work in Progress"
-staleLabel: 'Meta: Abandoned'
-markComment: >
-    This submission has been flagged as abandoned due to lack of activity. Is there something specific that is causing this to be delayed?
-
-    Please leave a comment if this issue has not yet been resolved, otherwise it will be automatically closed in the near future.
-closeComment: >
-    Following with the previous comment, this submission has been automatically closed.
-
-    This was due to the following:
-      - No activity has been made for the past 14 days.
-      - There are no labels that indicate this is a protected issue.
-
-    If this issue is still relevant and needs to be opened again in the future, please leave a comment elaborating upon the reason to continue with this issue.
