--- .github/workflows/compile.yml
@@ -1,29 +1,0 @@
-name: "Java Continuous Integration"
-
-on:
-    push:
-        branches: [main, staging]
-    pull_request:
-        branches: [main, staging]
-
-jobs:
-    validate:
-        runs-on: ubuntu-latest
-
-        name: Java Maven Build
-        steps:
-            - uses: actions/checkout@v3
-            - name: Set up JDK 17
-              uses: actions/setup-java@v3
-              with:
-                  java-version: '17'
-                  distribution: 'adopt'
-                  architecture: x64
-                  server-id: github
-                  server-username: GITHUB_USER_REF
-                  server-password: GITHUB_TOKEN_REF
-            - name: Compile
-              run: "mvn -B -Pgithub --update-snapshots clean package --file pom.xml"
-              env:
-                GITHUB_USER_REF: ${{ secrets.GH_USERNAME }}
-                GITHUB_TOKEN_REF: ${{ secrets.GH_PAT }}
