+++ .gitignore
@@ -7,6 +7,11 @@
 
 # IntelliJ
 out/
+# mpeltonen/sbt-idea plugin
+.idea_modules/
+
+# JIRA plugin
+atlassian-ide-plugin.xml
 
 # Compiled class file
 *.class
@@ -95,19 +100,19 @@ $RECYCLE.BIN/
 # Windows shortcuts
 *.lnk
 
-target/
+.gradle
+build/
 
-pom.xml.tag
-pom.xml.releaseBackup
-pom.xml.versionsBackup
-pom.xml.next
+# Ignore Gradle GUI config
+gradle-app.setting
 
-release.properties
-dependency-reduced-pom.xml
-buildNumber.properties
-.mvn/timing.properties
-.mvn/wrapper/maven-wrapper.jar
-.flattened-pom.xml
+# Cache of project
+.gradletasknamecache
+
+**/build/
 
 # Common working directory
 run/
+
+# Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
+!gradle-wrapper.jar
