+++ build.gradle
@@ -1,0 +1,47 @@
+plugins {
+    id 'com.gradleup.shadow' version '8.3.0'
+    id 'java'
+}
+
+group = 'github.xCykrix'
+version = '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+    maven {
+        name = "spigotmc-repo"
+        url = "https://hub.spigotmc.org/nexus/content/repositories/snapshots/"
+    }
+    maven {
+        name = "sonatype"
+        url = "https://oss.sonatype.org/content/groups/public/"
+    }
+    flatDir {
+        dirs("../SpigotDevkit/build/libs")
+    }
+}
+
+dependencies {
+    compileOnly("org.spigotmc:spigot-api:1.21-R0.1-SNAPSHOT")
+    implementation("github.xCykrix:SpigotDevkit:1.0-0-SNAPSHOT-all")
+}
+
+// Target Java Build (Java 17, Minecraft 1.17.x)
+def targetJavaVersion = 17
+java {
+    def javaVersion = JavaVersion.toVersion(targetJavaVersion)
+    sourceCompatibility = javaVersion
+    targetCompatibility = javaVersion
+    if (JavaVersion.current() < javaVersion) {
+        toolchain.languageVersion = JavaLanguageVersion.of(targetJavaVersion)
+    }
+}
+
+processResources {
+    def props = [version: version]
+    inputs.properties props
+    filteringCharset 'UTF-8'
+    filesMatching('plugin.yml') {
+        expand props
+    }
+}
